requirements.txt
numpy
scipy
tqdm
torch==1.8
torchvision==0.9
matplotlib
tensorboard
pandas
pyyaml
minio
pytorch-lightning
lightning-bolts
wget
scikit-image